Find TEXTDOMAINDIR relative to script
Previously, downgrade hard-coded /usr/local/share/locale. This meant
installation to an alternative PREFIX would still use /usr/local for the
locale files, which is surprising.

This change means that wherever downgrade is installed, it will find
locales relative to itself. This is behavior neutral on Arch, since the
PKGBUILD[1] uses a PREFIX of /usr/local.

A future commit will bring the locale files installation into the make
install target, which will respect PREFIX. That will only function with
the pre-requisite commit.