Fixup tests for new pacman-conf invocations
When we did that, we were using `make test` on CI instead of `just test`
and `make test` just quietly exits successfully ("nothing to do")
because a `test` directory exists. Nice.

I thought the fix is to record new fixtures and ensure we always write
valid configurations (everything in a section), so I did that first.

In actuality that is hard since the invocation includes a tempfile the
hashed command-name (for looking fixtures) changes run to run.
Ultimately, I chose to stub the read function for now. However, I kept
the "write a valid conf" fix since it's not a bad thing and may prove
useful in the future.