feat: add ability to downgrade aur package using git
This is fully opt-in by the `--git` option (or equivalent env var).

If enabled, we check to see if the configured cache directory represents
the git repository. This would be the case if the user is downgrading an
AUR package installed via `yay`. In such a repository, we can hunt for
commits that represent the older versions of the package and present
them as options to downgrade to. If selected, we can check out that
commit and build the package at that version (if necessary) and install
it.

The logic is isolated to a `search_git` function, for handling locating
commits that represent versions and output potential package names that
include "gitpkg" in them, and a `build_pkg` function that looks for such
package names and builds them. We have to be careful to use `su` so that
`makepkg` is not run as root. The only changes to `downgrade` outside of
these new functions is to regexs used to locate and parse package names,
but this is well covered by tests.